pyproject.toml
[project]
name = "wikisearch"
version = "0.1.0"
description = "Token-efficient search module for LlmBrain wiki"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "python-frontmatter>=1.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "fastmcp>=2.0",
]

[project.scripts]
wiki = "wikisearch.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wikisearch"]
